requirements.txt
absl-py==1.2.0
colorlog==6.6.0
gpytorch==1.8.0
hydra-colorlog==1.2.0
hydra-core==1.2.0
imageio==2.20.0
ipython==8.4.0
joblib==1.1.0
matplotlib==3.5.2
nflows==0.14
numpy==1.23.1
omegaconf==2.2.2
pandas==1.4.3
Pillow==9.2.0
pyro-ppl==1.8.1
pytorch-lightning==1.6.5
rich==12.5.1
scipy==1.8.1
sdeint==0.3.0
seaborn==0.11.2
sklearn==0.0
scikit-image
h5py
skorch==0.8.0
sympy==1.10.1
tensorboard==2.9.1
tqdm==4.64.0
wandb==0.12.21

